+++ src/parser/random.h
@@ -5,6 +5,6 @@
 
 void setRandSeed(int64_t randseed);
 int64_t genRandInt(int64_t l, int64_t r);
-long double genRandFloat(int64_t l, int64_t r, int d);
+long double genRandFloat(int64_t l, int64_t r, int64_t d);
 
 #endif // __RANDOM_H__
